+++ plone/app/viewletmanager/testing.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import IntegrationTesting
+from plone.app.testing import PLONE_FIXTURE
 from plone.testing import Layer
 from zope.configuration import xmlconfig